+++ projects/js-packages/ai-client/src/components/ai-control/extension-ai-control.tsx
@@ -46,6 +46,8 @@ type ExtensionAIControlProps = {
 	onUndo?: () => void;
 	onUpgrade?: ( event: MouseEvent< HTMLButtonElement > ) => void;
 	onTryAgain?: () => void;
+	lastAction?: string;
+	blockType: string;
 };
 
 /**
@@ -78,6 +80,8 @@ export function ExtensionAIControl(
 		onUndo,
 		onUpgrade,
 		onTryAgain,
+		lastAction,
+		blockType,
 	}: ExtensionAIControlProps,
 	ref: React.MutableRefObject< HTMLInputElement >
 ): ReactElement {
@@ -86,8 +90,6 @@ export function ExtensionAIControl(
 	const [ lastValue, setLastValue ] = useState( value || null );
 	const promptUserInputRef = useRef( null );
 	const isDone = value?.length <= 0 && state === 'done';
-	const [ initialPlaceholder ] = useState( placeholder );
-	const [ prompt, setPrompt ] = useState( null );
 
 	// Pass the ref to forwardRef.
 	useImperativeHandle( ref, () => promptUserInputRef.current );
@@ -98,16 +100,8 @@ export function ExtensionAIControl(
 		}
 	}, [ editRequest ] );
 
-	useEffect( () => {
-		if ( placeholder !== initialPlaceholder ) {
-			// The prompt is used to determine if there was a toolbar action
-			setPrompt( placeholder );
-		}
-	}, [ placeholder ] );
-
 	const sendHandler = useCallback( () => {
 		setLastValue( value );
-		setPrompt( value );
 		setEditRequest( false );
 		onSend?.( value );
 	}, [ onSend, value ] );
@@ -246,9 +240,12 @@ export function ExtensionAIControl(
 	} else if ( showGuideLine ) {
 		message = isDone ? (
 			<GuidelineMessage
-				showAIFeedbackThumbs={ true }
-				ratedItem={ 'ai-assistant' }
-				prompt={ prompt }
+				aiFeedbackThumbsOptions={ {
+					showAIFeedbackThumbs: true,
+					ratedItem: 'ai-assistant',
+					prompt: lastAction,
+					block: blockType,
+				} }
 			/>
 		) : (
 			<GuidelineMessage />